style: convert configs from js to ts
