+++ frontend/svelte.config.ts
